CI: split build jobs (#3300)
When we want to test a change we need the images therefore we should have a CI
that builds them reasonably quickly. Currently this leaves much to be desired
with build times of about half an hour. Changes

- Create separate jobs to build release binaries. This significantly speeds up
  the build.
- Configure caches to be saved on main and release branches and on
  workflow_dispatch for easier debugging.
- Switch to using ubuntu-latest: it's a bit faster on our custom 8CPU runner but
  we just end up having to wait for the arm build to complete.
- Add a docker base image so we don't have to apt-get and download the SRS
  during each docker build:
  https://github.com/EspressoSystems/docker-base-images/. These steps are very
  slow when building the ARM image with qemu.
- Remove caches from CI jobs that don't really need them.

With a full cache hit it take a bit less than 10 minutes now until all docker
images are built. Without the cache I think it takes about 17 minutes.

It took me a while but we do unfortunately need separate caches for some of the
jobs for max effectiveness. The sequencer-sqlite binary also has it's own
lockfile so many versions don't match. This is quite annoying but maybe tricky
to fix. I haven't looked into it. The log of cargo build in the cached run shows
that for all jobs only workspace crates are compiled, which means that the cache
hit is as good as possible with the default setting for the rust-cache action.