CI: native docker builds (#3344)
* CI: native docker builds

QEMU is very slow which limits us to what we can do in docker builds
within reasonable time. We do want images to build quickly in order to
deploy them. This PR builds native images on AMD and ARM architectures
and then merges them into a multiplatform image.

* fix Dockerfile path

* fix: separate digests for each service

* CI: try public arm runners

* fix runner label

* remove buildjet cache

* re-enable branch gating

* update comment

* CI: add the tag on PR builds

* aggregate test demo jobs

* fix job dependency

* add architecture to job name

Makes it easier to click on the right thing in the github UI