Add validators and stake table to data state
In order for the UI of the Block Explorer's Node Page to be able to
surface relevant information related to the validators and their ability
to delegate, we need to be able to send the relevant information. This
information includes the total list of validators, and the active validator
set.

The active validator set is ultimately determined by the active stake
table.  While this information is actionable by itself, it is an incomplete
picture as the resulting validator set would only consist of those entries
in the stake table.  We try to ensure that we don't drop any existing
validators, by only adding to the list, replacing existing entries, or
adding missing ones.

The information about the current stake table, and validator list can
be subscribed and queried by the websocket users.

This commit adds the ability to request and retrieve information about
the current stake table, and the list of validators.

NOTE:  There is still some work to be done here in order to make this
entirely effective and seamless.
- This will need to consume a stream of the Events to add validators
and add delegations in order to keep the **FULL** list of validators up-to-date
- This will need to submit new stake table and validator events to let
the consuming user know that there is new information available