Distinguish between namespace indexes vs IDs (#3361)
* Distinguish between namespace indexes vs IDs

The earlier change to separate namespace information in transactions
in the database was incorrect, because it conflated namespace indexes
(i.e. the offset of a given namespace within the namespace table,
which may vary from block to block for a given namespace) with global
namespace IDs.

I discovered this when I tried to implement transaction filtering by
namespace for the explorer, which should have been easy after the
previous change. It turned out not to be possible, because what was
stored in the database was only the namespace index, not the ID.

This change addresses the problem and hopefully reduces confusion
by creating separate types for namespace index vs ID. Both are now
stored in the database (the migration to populate the namespace ID
is a little gross, but thankfully the information is all there in
the namespace table).

We add a test for the new functionality, and implement the explorer
namespace filtering feature as a sanity check that the database schema
is now correct and sufficently expressive. Thus, in addition to running
unit tests, you can test this change by running a local demo, opening
the block explorer in a browser, and clicking on a namespace link in
the transactions panel. This will take you to a page listing transactions
in that namespace, which should be correctly populated. Prior to this
change, that page was empty.

Incidentally, this change also fixes what I believe is a bug: when
building a block, bytes added to the namespace table are counted
against the total block size for the purpose of truncating, but
these bytes actually live in the header, and don't count when
enforcing maximum block size in the actual proposal validation. I
discovered this while working on the new unit test, since that also
deals with transaction and namespace sizes.

Note that the updated migration for Postgres populates all the new
fields using a coputation on existing data, but the sqlite migration
creates a new _empty_ table. This is basically because sqlite doesn't
have a base64 decode function, which is needed to get the namespace ID
out of the namespace table. There are extensions for this, but there is
not a good package manager, so it's difficult to install and manage these
extensions in a nice way. For now, I think it's ok for the migration to
adjust the schema but not populate old data, since we don't have any sqlite
nodes in production anyway.

* Fix postgres transactions migration and add regression test